CustomTrainer/requirements.txt
pillow>=10.0.0
PyYAML>=6.0
ultralytics>=8.2.0
opencv-python>=4.8.0
numpy>=1.24,<3
